refactor : 필요없는 내용 제거 및 폴더 이동(3)
