+++ src/main/java/org/gdg/zipte_gdg/security/jwt/filter/TokenAuthenticationFilter.java
@@ -7,7 +7,7 @@ import jakarta.servlet.http.HttpServletRequest;
 import jakarta.servlet.http.HttpServletResponse;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.log4j.Log4j2;
-import org.gdg.zipte_gdg.security.jwt.util.TokenProvider;
+import org.gdg.zipte_gdg.security.jwt.provider.TokenProvider;
 import org.springframework.security.core.Authentication;
 import org.springframework.security.core.context.SecurityContextHolder;
 import org.springframework.stereotype.Component;
